+++ device-examples/msp430fr2355/examples/lpm0.rs
@@ -20,20 +20,6 @@ use panic_msp430 as _;
 
 static P2IV: Mutex<RefCell<Option< PxIV<P2> >>> = Mutex::new(RefCell::new(None));
 
-macro_rules! init_port_as_pulldowns {
-    ($port: expr) => {
-        Batch::new($port)
-            .config_pin0(|p| p.pulldown())
-            .config_pin1(|p| p.pulldown())
-            .config_pin2(|p| p.pulldown())
-            .config_pin3(|p| p.pulldown())
-            .config_pin4(|p| p.pulldown())
-            .config_pin5(|p| p.pulldown())
-            .config_pin6(|p| p.pulldown())
-            .config_pin7(|p| p.pulldown())
-    };
-}
-
 // P1.0 should toggle when P2.3 is pressed
 #[entry]
 fn main() -> ! {
@@ -44,12 +30,14 @@ fn main() -> ! {
 
     // Floating input pins consume a *huge* amount of power (relatively speaking).
     // Set unused pins to outputs or enable their pull resistors.
-    let p1 = init_port_as_pulldowns!(periph.p1)
+    let p1 = Batch::new(periph.p1)
+        .pulldown_all()
         .config_pin0(|p| p.to_output())
         .split(&pmm);
     let mut red_led = p1.pin0;
 
-    let p2 = init_port_as_pulldowns!(periph.p2)
+    let p2 = Batch::new(periph.p2)
+        .pulldown_all()
         .config_pin3(|p| p.pullup())
         .split(&pmm);
     let mut button = p2.pin3;
@@ -89,10 +77,10 @@ fn PORT2() {
 
 /// Enable pulldowns on unused ports to massively reduce power usage.
 fn init_unused_gpio(p3: P3, p4: P4, p5: P5, p6: P6, pmm: &Pmm) {
-    init_port_as_pulldowns!(p3).split(pmm);
-    init_port_as_pulldowns!(p4).split(pmm);
-    init_port_as_pulldowns!(p5).split(pmm);
-    init_port_as_pulldowns!(p6).split(pmm);
+    Batch::new(p3).pulldown_all().split(pmm);
+    Batch::new(p4).pulldown_all().split(pmm);
+    Batch::new(p5).pulldown_all().split(pmm);
+    Batch::new(p6).pulldown_all().split(pmm);
 }
 
 // The compiler will emit calls to the abort() compiler intrinsic if debug assertions are